pyproject.toml
[tool.poetry]
name = "lux-beauty"
version = "0.1.0"
description = "Legally compliant scraping pipeline for French luxury beauty e-commerce data"
authors = ["Research Team <research@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
playwright = "^1.40.0"
httpx = "^0.25.0"
aiohttp = "^3.9.0"
requests = "^2.31.0"
lxml = "^4.9.0"
beautifulsoup4 = "^4.12.0"
selectolax = "^0.3.0"
pydantic = "^2.5.0"
pandas = "^2.1.0"
polars = "^0.20.0"
duckdb = "^0.9.0"
pyarrow = "^14.0.0"
structlog = "^23.2.0"
pyyaml = "^6.0.1"
click = "^8.1.0"
    langdetect = "^1.0.9"
        great-expectations = "^0.18.0"
    scikit-learn = "^1.4.0"
    pandera = "^0.17.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-playwright = "^0.4.0"
responses = "^0.24.0"
ruff = "^0.1.0"
black = "^23.11.0"
mypy = "^1.7.0"
pyright = "^1.1.0"

[tool.poetry.scripts]
lux-beauty = "src.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
